Implemented new linker

+++ Source/Mosa.Kernel.x86/CMOS.cs
@@ -26,7 +26,7 @@ namespace Mosa.Kernel.x86
 			Native.Nop();
 			Native.Nop();
 			Native.Nop();
-			byte result = (byte)Native.In(0x71);
+			byte result = Native.In(0x71);
 
 			//Native.Sti();
 			return result;
